only bust browser cache when assets themselves change
This cache-busting strategy is more realistic and less aggressive, because
it only forces the browser to download a new asset bundle when an asset in
that bundle changes. For example, using the standard assets.version/timestamp
strategy, every time the CSS changes, all the JS will also be re-downloaded.
With this new strategy, when the CSS changes, only the CSS gets
re-downloaded.